% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aIc.plot.R
\name{aIc.plot}
\alias{aIc.plot}
\title{\code{aIc.plot} plots the result of the distance tests.}
\usage{
aIc.plot(test.out)
}
\arguments{
\item{test.out}{is the output from either aIc.dominant, aIc.scale, aIc.perturb}
}
\value{
returns a plot of the density of the distance test results.
  test result.
}
\description{
\code{aIc.plot} plots the result of the distance tests.
}
\examples{
data(selex)
group = c(rep('N', 7), rep('S', 7))
test.out <- aIc.dominant(selex, norm.method='prop', group=group)
aIc.plot(test.out)
}
\author{
Greg Gloor
}
