% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aIc.runExample.r
\name{aIc.runExample}
\alias{aIc.runExample}
\title{\code{aIc.runExample} loads the associated shiny app}
\usage{
aIc.runExample()
}
\description{
This will load the selex example dataset with the default group sizes,
the user can upload their own local dataset and adjust groups accordingly.
}
\examples{
library(aIc)
aIc.runExample()
}
\author{
Greg Gloor
}
