% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{meta16S}
\alias{meta16S}
\title{16S rRNA tag-sequencing data}
\format{
A data frame with 359 columns and 860 rows
}
\source{
doi: 10.1128/mSphere.00327-17
}
\usage{
data(meta16S)
}
\description{
A count table of a 16S rRNA amplicon data
Two groups, pupils and centenarians are represented with 198
and 161 samples per group respectively. samples are by column  
and OTU ids are by row.
}
\keyword{datasets}
