% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{metaTscome}
\alias{metaTscome}
\title{meta-transcriptome data}
\format{
A data frame with 17 columns and 3647 rows
}
\source{
doi:10.1007/978-3-030-71175-7_17 and doi:10.1007/978-1-4939-8728-3_13
}
\usage{
data(metaTscome)
}
\description{
A count table of a mixed population or metatranscriptome 
experiment. Two groups, H and BV are represented with 7 and 10
samples per group respectively. samples are by column and 
functions are by row.
}
\keyword{datasets}
