% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{selex}
\alias{selex}
\title{Selection-based differential sequence variant abundance dataset}
\format{
A data frame with 14 columns and 1600 rows
}
\source{
DOI:10.1073/pnas.1322352111
}
\usage{
data(selex)
}
\description{
This data set gives the differential abundance of 1600 enzyme variants
grown under selective (NS) and selective (S) conditions
}
\keyword{datasets}
