% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{singleCell}
\alias{singleCell}
\title{single cell transcriptome data}
\format{
A data frame with 2000 columns and 1508 rows
}
\source{
https://www.nature.com/articles/s41592-019-0372-4
}
\usage{
data(singleCell)
}
\description{
A count table of a single cell transcriptome data subset from 
the count table from  doi:10.1038/s41592-019-0372-4. Two groups
memory T cells, and cytotoxic T cells, 1000 cells per group.
samples are by column and genes are by row.
}
\keyword{datasets}
