% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{transcriptome}
\alias{transcriptome}
\title{Saccharomyces cerevisiae transcriptome}
\format{
A data frame with 96 columns and 5892 rows
}
\source{
DOI: 10.1261/rna.053959.115 and PRJEB5348
}
\usage{
data(transcriptome)
}
\description{
A count table of a highly replicated RNA-seq experiment
with samples by column and genes by row. Two groups composed
of SNF2 knockout and WT, 48 samples in each.
}
\keyword{datasets}
